[DX] Extract JobInstance
