+++ spec/Akeneo/Bundle/BatchBundle/Job/JobInstanceFactorySpec.php
@@ -6,7 +6,7 @@ use PhpSpec\ObjectBehavior;
 
 class JobInstanceFactorySpec extends ObjectBehavior
 {
-    const TESTED_CLASS = 'Akeneo\Bundle\BatchBundle\Entity\JobInstance';
+    const TESTED_CLASS = 'Akeneo\Component\Batch\Model\JobInstance';
 
     function let()
     {
