+++ spec/Pim/Bundle/BaseConnectorBundle/Archiver/ArchivableFileWriterArchiverSpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Bundle\BaseConnectorBundle\Archiver;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
-use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
+use Akeneo\Component\Batch\Model\JobInstance;
 use Akeneo\Component\Batch\Item\ItemWriterInterface;
 use Akeneo\Component\Batch\Job\Job;
 use Akeneo\Component\Batch\Step\AbstractStep;
