+++ spec/Pim/Bundle/VersioningBundle/EventSubscriber/AddContextSubscriberSpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Bundle\VersioningBundle\EventSubscriber;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
-use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
+use Akeneo\Component\Batch\Model\JobInstance;
 use Akeneo\Component\Batch\Event\EventInterface;
 use Akeneo\Component\Batch\Event\JobExecutionEvent;
 use PhpSpec\ObjectBehavior;
