+++ src/Akeneo/Bundle/BatchBundle/Launcher/SimpleJobLauncher.php
@@ -3,7 +3,7 @@
 namespace Akeneo\Bundle\BatchBundle\Launcher;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
-use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
+use Akeneo\Component\Batch\Model\JobInstance;
 use Akeneo\Component\Batch\Job\JobRepositoryInterface;
 use Symfony\Component\Process\PhpExecutableFinder;
 use Symfony\Component\Security\Core\User\UserInterface;
