+++ src/Pim/Bundle/ImportExportBundle/Entity/Repository/JobInstanceRepository.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\ImportExportBundle\Entity\Repository;
 
-use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
+use Akeneo\Component\Batch\Model\JobInstance;
 use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Doctrine\ORM\EntityRepository;
