+++ src/Pim/Bundle/ImportExportBundle/Resources/config/validation/jobinstance.yml
@@ -1,4 +1,4 @@
-Akeneo\Bundle\BatchBundle\Entity\JobInstance:
+Akeneo\Component\Batch\Model\JobInstance:
     constraints:
         - Pim\Bundle\ImportExportBundle\Validator\Constraints\JobInstance:
             property: alias
