+++ src/Pim/Bundle/ImportExportBundle/Resources/views/JobProfile/_show_headers.html.twig
@@ -9,7 +9,7 @@
 
 {% set buttons %}
     {{ elements.backLink(indexRoute) }}
-    {% if jobInstance.status != constant('Akeneo\\Bundle\\BatchBundle\\Entity\\JobInstance::STATUS_IN_PROGRESS') %}
+    {% if jobInstance.status != constant('Akeneo\\Component\\Batch\\Model\\JobInstance::STATUS_IN_PROGRESS') %}
         {{ elements.link('Edit', editRoute, { icon: 'pencil' }) }}
     {% endif %}
     {% if uploadAllowed and uploadViolations|length == 0 %}
