+++ src/Pim/Bundle/InstallerBundle/DataFixtures/AbstractLoadFixturesData.php
@@ -4,7 +4,7 @@ namespace Pim\Bundle\InstallerBundle\DataFixtures;
 
 use Akeneo\Bundle\BatchBundle\Command\BatchCommand;
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
-use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
+use Akeneo\Component\Batch\Model\JobInstance;
 use Akeneo\Component\Batch\Job\ExitStatus;
 use Doctrine\Common\DataFixtures\AbstractFixture;
 use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
