+++ src/Pim/Bundle/TransformBundle/Normalizer/Flat/JobInstanceNormalizer.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\TransformBundle\Normalizer\Flat;
 
-use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
+use Akeneo\Component\Batch\Model\JobInstance;
 use Pim\Bundle\TransformBundle\Normalizer\Structured;
 
 /**
