+++ src/Pim/Bundle/VersioningBundle/EventSubscriber/AddContextSubscriber.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\VersioningBundle\EventSubscriber;
 
-use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
+use Akeneo\Component\Batch\Model\JobInstance;
 use Akeneo\Component\Batch\Event\EventInterface;
 use Akeneo\Component\Batch\Event\JobExecutionEvent;
 use Pim\Bundle\VersioningBundle\Manager\VersionContext;
